+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [1.9.2-next.3](https://github.com/warp-ds/css/compare/v1.9.2-next.2...v1.9.2-next.3) (2024-03-28)
+
+
+### Bug Fixes
+
+* Update warp-ds/uno ([#188](https://github.com/warp-ds/css/issues/188)) ([20909d2](https://github.com/warp-ds/css/commit/20909d274cbdcb7bf5eb0b04165aed78459738f7))
+
+## [1.9.2-next.2](https://github.com/warp-ds/css/compare/v1.9.2-next.1...v1.9.2-next.2) (2024-03-28)
+
+
+### Bug Fixes
+
+* Resolved underline issues in Button component ([#187](https://github.com/warp-ds/css/issues/187)) ([4812cfd](https://github.com/warp-ds/css/commit/4812cfd224de2709f7b26faab6293f5dc18efc43))
+
+## [1.9.2-next.1](https://github.com/warp-ds/css/compare/v1.9.1...v1.9.2-next.1) (2024-03-27)
+
+
+### Bug Fixes
+
+* Update Box and Card to use new surface concept ([#186](https://github.com/warp-ds/css/issues/186)) ([215d8cf](https://github.com/warp-ds/css/commit/215d8cf37287f146da4c66a38725fe1a80b6c17f))
+
 ## [1.9.1](https://github.com/warp-ds/css/compare/v1.9.0...v1.9.1) (2024-03-20)